fixed drive motors & inverted them *read desc*, lift can't go down if more than 1 ball in it, made toggleable far shooting*
adjusted for CTRE's master/slave code error, then inverted the driving to make collector forward for autos. We have to change it back later for teleop though.

toggleable far shooting meant I had to make a second set of data points for the interpolating tree map in Constants and adjusting the shoot command